update memory monitor
